Fix payload extraction time window
